% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{accuracy_tibble}
\alias{accuracy_tibble}
\title{Compute accuracy based on classificaiton table.}
\usage{
accuracy_tibble(class_dt)
}
\arguments{
\item{class_dt}{output of sn_classifier.}
}
\value{
an accuracy score with upper and lower bounds.
}
\description{
Compute accuracy based on classificaiton table.
}
