% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data.R
\name{build_model_data}
\alias{build_model_data}
\title{Build data to be used for model fitting}
\usage{
build_model_data(sscorpus, states = NULL, min_states = 1)
}
\arguments{
\item{sscorpus}{sscorpus object.}

\item{states}{character vector of tokens to be included in the model.}

\item{min_states}{an integer specifying a minimum semantic fluency list length.}
}
\value{
a list with model parameters.
}
\description{
Build data to be used for model fitting
}
