% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sn.R
\name{estimate_sn}
\alias{estimate_sn}
\title{Estimate semantic network}
\usage{
estimate_sn(sscorpus, cue_set, tag_set, num_states = 30,
  num_folds = 10, tags_AND = TRUE, remove_cue = TRUE,
  min_count = NULL, min_states = 1, states = NULL)
}
\arguments{
\item{sscorpus}{sscorpus object.}

\item{tag_set}{character vector containing the tags characterizing the groups of interest.}

\item{num_states}{ingeger value specifying number of top tokens.}

\item{num_folds}{ingeger value specifying number of folds for cross-validation.}

\item{tags_AND}{logical function specifying whether tags apply with an AND or an OR condition (e.g. "republican" and/or "male").}

\item{remove_cue}{logical value specifying whether cue should be removed from the top of every list.}

\item{min_count}{integer value specifying if only tokens with a given minimum count should be selected.}

\item{min_states}{an integer specifying a minimum semantic fluency list length.}

\item{states}{character vector specifying tokens (states) to be included in model.}

\item{cue}{character indicating cue.}
}
\value{
list of length 2: (a) a fitted model (b) paramaters of a fitted model (transition matrix and initial probabilities).
}
\description{
Estimate semantic network
}
