% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fit.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters from a fitted stan model.}
\usage{
extract_parameters(fit, states)
}
\arguments{
\item{fit}{fitted model.}

\item{states}{character vector of tokens (states).}
}
\value{
list of length 2: (a) pi = numeric vector of initial probabilities (b) P = numeric matrix of transition probabilities.
}
\description{
Extract parameters from a fitted stan model.
}
