% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data.R
\name{fluency_to_int}
\alias{fluency_to_int}
\title{Convert lists of tokens into lists of integers using a pre-defined index (output of state_to_int).}
\usage{
fluency_to_int(sscorpus, states_index, min_states = 1)
}
\arguments{
\item{sscorpus}{an sscorpus (generally a subsetted sscorpus)}

\item{states_index}{a list of 2: (1) the original character vector (2) a vector of integers, one for each token. Output of state_to_int.}

\item{min_states}{an integer specifying a minimum semantic fluency list length.}
}
\value{
a data.frame with 3 cols: (1) pid = subject id (2) state_ID = fluency lists in integer form (3) num_states = # of states visited
}
\description{
Convert lists of tokens into lists of integers using a pre-defined index (output of state_to_int).
}
