% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{ll_table}
\alias{ll_table}
\title{Build classification table given model fits}
\usage{
ll_table(cue, fits, test_sets)
}
\arguments{
\item{cue}{character indicating cue.}

\item{fits}{list of model fits of length (num_folds)}

\item{test_sets}{list of test-sets, one for each model fit.}
}
\value{
data.frame with each subject's log-likelihood under each model, a classification and a true label.
}
\description{
Build classification table given model fits
}
