% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data.R
\name{pad_fluency}
\alias{pad_fluency}
\title{Pad fluency lists}
\usage{
pad_fluency(state_ID, states_index)
}
\arguments{
\item{state_ID}{fluency lists in integer form (column of fluency_to_int output).}

\item{states_index}{a list of 2 (output of state_to_int): (1) the original character vector (2) a vector of integers, one for each token.}
}
\value{
a 1 column data.frame with fluency lists in wide integer form (all lists of equal length).
}
\description{
Pad fluency lists
}
