% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data_cv.R
\name{partition_data_cv}
\alias{partition_data_cv}
\title{Partition data into training and test sets for cross-validation.}
\usage{
partition_data_cv(model_data, num_folds = 10)
}
\arguments{
\item{model_data}{model_data object (output of build_model_data).}

\item{num_folds}{integer value specifying number of folds for cross-validation.}
}
\value{
list of num_fold training sets each with its respective test set.
}
\description{
Partition data into training and test sets for cross-validation.
}
