% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{sn_classifier}
\alias{sn_classifier}
\title{Train a semantic-network based classifier}
\usage{
sn_classifier(cue_set, tags, sscorpus, num_states = 30, num_folds = 10,
  tags_AND = TRUE, remove_cue = TRUE, min_count = NULL,
  min_states = 1)
}
\arguments{
\item{cue_set}{character indicating cue.}

\item{tags}{character vector or a list of length 2 containing the tags characterizing the two groups of interest.}

\item{sscorpus}{sscorpus object.}

\item{num_states}{ingeger value specifying number of top tokens.}

\item{num_folds}{ingeger value specifying number of folds for cross-validation.}

\item{tags_AND}{logical function specifying whether tags apply with an AND or an OR condition (e.g. "republican" and/or "male").}

\item{remove_cue}{logical value specifying whether cue should be removed from the top of every list.}

\item{min_count}{integer value specifying if only tokens with a given minimum count should be selected.}

\item{min_states}{an integer specifying a minimum semantic fluency list length.}
}
\value{
data partitioned into 10 training sets each with its respective test fold.
}
\description{
Train a semantic-network based classifier
}
