% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data.R
\name{state_to_int}
\alias{state_to_int}
\title{Assign a unique integer to each token (state).}
\usage{
state_to_int(states)
}
\arguments{
\item{states}{a character vector of tokens.}
}
\value{
a list of 2: (1) the original character vector (2) a vector of integers, one for each token.
}
\description{
Assign a unique integer to each token (state).
}
