% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{subset_data}
\alias{subset_data}
\title{Subset model_data provided a set of ids.}
\usage{
subset_data(ids, model_data)
}
\arguments{
\item{ids}{subject ids.}

\item{model_data}{model_data object (output of build_model_data).}
}
\value{
subsetted model_data object.
}
\description{
Subset model_data provided a set of ids.
}
