% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data_cv.R
\name{subset_data_cv}
\alias{subset_data_cv}
\title{Subset model_data provided a set of ids for a training and a test set.}
\usage{
subset_data_cv(ids, model_data)
}
\arguments{
\item{ids}{subject ids partitioned into training and test set (one fold of subset_ids_cv output).}

\item{model_data}{model_data object (output of build_model_data).}
}
\value{
model_data object for training and test-set.
}
\description{
Subset model_data provided a set of ids for a training and a test set.
}
