% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{subset_ids}
\alias{subset_ids}
\title{Partition subject ids into training and test set.}
\usage{
subset_ids(model_data, prop_train = 0.9)
}
\arguments{
\item{model_data}{model_data object (output of build_model_data).}

\item{prop_train}{proportion of observations to be used in training set.}
}
\value{
subject ids partitioned into a training- and a test-set.
}
\description{
Partition subject ids into training and test set.
}
