% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data_cv.R
\name{subset_ids_cv}
\alias{subset_ids_cv}
\title{Partition subject ids into num_fold folds}
\usage{
subset_ids_cv(model_data, num_folds = 10)
}
\arguments{
\item{model_data}{model_data object (output of build_model_data).}

\item{num_folds}{integer value specifying number of folds for cross-validation.}
}
\value{
subject ids partitioned into num_fold training sets each with its respective test set.
}
\description{
Partition subject ids into num_fold folds
}
