% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sscorpus.R
\name{subset_sscorpus}
\alias{subset_sscorpus}
\title{Subset and sscorpus by cue and tags.}
\usage{
subset_sscorpus(sscorpus, cue_set = NULL, tag_set = NULL,
  tags_AND = TRUE)
}
\arguments{
\item{sscorpus}{sscorpus object.}

\item{cue_set}{the cues of interest.}

\item{tag_set}{the tags that characterize the subjects of interest.}

\item{tags_AND}{logical function specifying whether tags apply with an AND or an OR condition (e.g. "republican" and/or "male").}
}
\value{
an sscorpus (subsetted if specified).
}
\description{
Subset and sscorpus by cue and tags.
}
