% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sscorpus.R
\name{subset_sscorpus_ids}
\alias{subset_sscorpus_ids}
\title{Subset and sscorpus by ids (used in classifier).}
\usage{
subset_sscorpus_ids(sscorpus, ids)
}
\arguments{
\item{sscorpus}{sscorpus object.}

\item{ids}{chracter vector of pids.}
}
\value{
an sscorpus (subsetted if specified).
}
\description{
Subset and sscorpus by ids (used in classifier).
}
