% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_states.R
\name{top_states}
\alias{top_states}
\title{Identify set of top tokens based on counts.}
\usage{
top_states(sscorpus, N = NULL, cue_set = NULL, tag_set = NULL,
  tags_AND = TRUE, remove_cue = TRUE, min_count = NULL)
}
\arguments{
\item{sscorpus}{sscorpus object.}

\item{N}{the number of top tokens to identify.}

\item{cue_set}{the cues of interest.}

\item{tag_set}{the tags that characterize the subjects of interest.}

\item{tags_AND}{logical function specifying whether tags apply with an AND or an OR condition (e.g. "republican" and/or "male").}

\item{remove_cue}{logical value specifying whether cue should be removed from the top of every list.}

\item{min_count}{integer value specifying if only states with a given minimum count should be selected.}

\item{min_states}{an integer specifying a minimum semantic fluency list length.}
}
\value{
a data.frame with 2 columns: (1) fluency = token (2) n = count.
}
\description{
Identify set of top tokens based on counts.
}
