% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data.R
\name{transit_matrix}
\alias{transit_matrix}
\title{Build an indexed transition matrix}
\usage{
transit_matrix(num_states)
}
\arguments{
\item{num_states}{an intenger specifying number of states}
}
\value{
a symmetric matrix with unique indices on each off-diagonal entry (diagonal is always 0 as self-loops are not allowed)
}
\description{
Build an indexed transition matrix
}
