% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{visualize_diff}
\alias{visualize_diff}
\title{Visualize semantic network difference (difference in initial and transition probabilities given parameters for two groups.)}
\usage{
visualize_diff(model_parameters, groups, mode = "undirected",
  threshold = NULL, edge.width.mult = 8, node.size.mult = 115,
  node.size.sum = 0.03)
}
\arguments{
\item{model_parameters}{list of length 2: (a) pi = numeric vector of initial probabilities (b) P = numeric matrix of transition probabilities.}

\item{mode}{whether network is "directed"/"undirected" (note, models are directed but using undirected can help visualization).}

\item{threshold}{a threshold applied to edges (edges below the threshold are left transparent).}

\item{edge.width.mult}{a multiplier to increase the width of edges.}

\item{node.size.mult}{a multiplier of node.size to greater differentiate between different initial probabilities.}

\item{node.size.sum}{a constant added to node.size to increase the baseline size of nodes.}
}
\value{
visualization of semantic network.
}
\description{
Visualize semantic network difference (difference in initial and transition probabilities given parameters for two groups.)
}
